pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-secrets"
version = "0.1.0"
description = "SOPS wrapper for the AI agent era - manage secrets without leaking them to LLM context"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "cptfinch" }
]
keywords = ["secrets", "sops", "age", "llm", "ai", "agent", "security", "encryption"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
llm-secrets = "llm_secrets.cli:main"
llms = "llm_secrets.cli:main"

[project.urls]
Documentation = "https://github.com/cptfinch/llm-secrets#readme"
Source = "https://github.com/cptfinch/llm-secrets"
Issues = "https://github.com/cptfinch/llm-secrets/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_secrets"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
]
